Object-Oriented-Programming-with-Java-pt1-University-of Helsinki-moocfi/week3-064.AverageOfNumbers/src/AverageOfNumbers.java

import java.util.ArrayList;

public class AverageOfNumbers {

    // Copy here the method sum from previous assignment
    public static int sum(ArrayList<Integer> list) {
        int total = 0;
        for(Integer i: list){total+=i;}
        return total;
    }
    

    public static double average(ArrayList<Integer> list) {
        return (double)sum(list)/(double)list.size();
    }

    public static void main(String[] args) {
        ArrayList<Integer> list = new ArrayList<Integer>();
        list.add(3);
        list.add(2);
        list.add(7);
        list.add(2);

        System.out.println("The average is: " + average(list));
    }
}
